+++ ofdb-application/src/lib.rs
@@ -9,6 +9,7 @@ mod create_event;
 mod create_place;
 mod create_rating;
 mod reset_password;
+mod review_place_with_token;
 mod review_places;
 mod send_update_reminders;
 mod update_event;
@@ -17,8 +18,9 @@ mod update_place;
 pub mod prelude {
     pub use super::{
         archive_comments::*, archive_events::*, archive_ratings::*, change_user_role::*,
-        create_event::*, create_place::*, create_rating::*, reset_password::*, review_places::*,
-        send_update_reminders::*, update_event::*, update_place::*,
+        create_event::*, create_place::*, create_rating::*, reset_password::*,
+        review_place_with_token::*, review_places::*, send_update_reminders::*, update_event::*,
+        update_place::*,
     };
 }