+++ ofdb-entities/src/url.rs
@@ -38,7 +38,7 @@ impl FromStr for Url {
 
 impl fmt::Display for Url {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self)
+        write!(f, "{}", self.0)
     }
 }